requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14
djangorestframework-simplejwt==4.7.2
httpie==3.2.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
pillow==10.3.0
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
daphne==4.1.2

